@@ -1,14 +1,14 @@
 package modelv2
 
 type Account struct {
-	Name          string                 `json:"name"`
-	Role          string                 `json:"role"`
-	Options       map[string]interface{} `json:"options"`
-	Secret        map[string]interface{} `json:"secret"`
-	Metadata      map[string]interface{} `json:"metadata"`
-	UserID        string                 `json:"userId"`
-	SecretVersion int                    `json:"secretVersion"`
-	Category      AccountCategory        `json:"rudderCategory"`
+	Name          string          `json:"name"`
+	Role          string          `json:"role"`
+	Options       map[string]any  `json:"options"`
+	Secret        map[string]any  `json:"secret"`
+	Metadata      map[string]any  `json:"metadata"`
+	UserID        string          `json:"userId"`
+	SecretVersion int             `json:"secretVersion"`
+	Category      AccountCategory `json:"rudderCategory"`
 }
 
 type AccountCategory string
