@@ -1,7 +1,7 @@
 package modelv2
 
 type Resource struct {
-	Name   string                 `json:"name"`
-	Role   string                 `json:"role"`
-	Config map[string]interface{} `json:"config"`
+	Name   string         `json:"name"`
+	Role   string         `json:"role"`
+	Config map[string]any `json:"config"`
 }
