@@ -11,8 +11,8 @@ type DataRetention struct {
 }
 
 type Bucket struct {
-	Type   string                 `json:"type"`
-	Config map[string]interface{} `json:"config"`
+	Type   string         `json:"type"`
+	Config map[string]any `json:"config"`
 }
 
 type Settings struct {
